+++ config/installer.php
@@ -1,0 +1,5 @@
+<?php
+
+return [
+    'locked' => env('APP_INSTALLED', false),
+];